+++ src/components/topbar/TopMenubar.vue
@@ -7,37 +7,7 @@
       :class="{ dropzone: isDropZone, 'dropzone-active': isDroppable }"
     >
       <h1 class="comfyui-logo mx-2">ComfyUI</h1>
-      <Menubar
-        :model="items"
-        class="top-menubar border-none p-0 bg-transparent"
-        :pt="{
-          rootList: 'gap-0 flex-nowrap w-auto',
-          submenu: `dropdown-direction-${dropdownDirection}`,
-          item: 'relative'
-        }"
-      >
-        <template #item="{ item, props, root }">
-          <a v-ripple class="p-menubar-item-link" v-bind="props.action">
-            <span
-              v-if="item.icon"
-              class="p-menubar-item-icon"
-              :class="item.icon"
-            />
-            <span class="p-menubar-item-label">{{ item.label }}</span>
-            <Badge
-              v-if="item.badge"
-              :class="{ 'ml-auto': !root, 'ml-2': root }"
-              :value="item.badge"
-            />
-            <span
-              v-if="!root && keybindings[item.id]"
-              class="ml-auto border border-surface rounded text-muted text-xs p-1 keybinding-tag"
-            >
-              {{ keybindings[item.id] }}
-            </span>
-          </a>
-        </template>
-      </Menubar>
+      <CommandMenubar />
       <Divider layout="vertical" class="mx-2" />
       <div class="flex-grow">
         <WorkflowTabs v-if="workflowTabsPosition === 'Topbar'" />
@@ -49,16 +19,14 @@
 </template>
 
 <script setup lang="ts">
-import Menubar from 'primevue/menubar'
 import Divider from 'primevue/divider'
 import WorkflowTabs from '@/components/topbar/WorkflowTabs.vue'
+import CommandMenubar from '@/components/topbar/CommandMenubar.vue'
 import Actionbar from '@/components/actionbar/ComfyActionbar.vue'
-import { ComfyMenuItem, useMenuItemStore } from '@/stores/menuItemStore'
 import { computed, onMounted, provide, ref } from 'vue'
 import { useSettingStore } from '@/stores/settingStore'
 import { app } from '@/scripts/app'
 import { useEventBus } from '@vueuse/core'
-import { useKeybindingStore } from '@/stores/keybindingStore'
 
 const settingStore = useSettingStore()
 const workflowTabsPosition = computed(() =>
@@ -72,32 +40,6 @@ const teleportTarget = computed(() =>
     ? '.comfyui-body-top'
     : '.comfyui-body-bottom'
 )
-const dropdownDirection = computed(() =>
-  settingStore.get('Comfy.UseNewMenu') === 'Top' ? 'down' : 'up'
-)
-
-const menuItemsStore = useMenuItemStore()
-const items = menuItemsStore.menuItems
-
-const keybindingStore = useKeybindingStore()
-const keybindings = computed(() => {
-  const bindings: Record<string, string> = {}
-  const stack: ComfyMenuItem[] = [...items]
-
-  while (stack.length) {
-    const item = stack.pop()!
-    if (item.id) {
-      const keybinding = keybindingStore.getKeybindingByCommandId(item.id)
-      if (keybinding) {
-        bindings[item.id] = keybinding.combo.toString()
-      }
-    }
-    if (item.items) {
-      stack.push(...item.items)
-    }
-  }
-  return bindings
-})
 
 const menuRight = ref<HTMLDivElement | null>(null)
 // Menu-right holds legacy topbar elements attached by custom scripts
@@ -146,21 +88,5 @@ eventBus.on((event: string, payload: any) => {
   font-size: 1.2em;
   user-select: none;
   cursor: default;
-}
-
-.keybinding-tag {
-  background: var(--p-content-hover-background);
-  border-color: var(--p-content-border-color);
-  border-style: solid;
-}
-</style>
-
-<style>
-.top-menubar .p-menubar-item-link svg {
-  display: none;
-}
-
-.p-menubar-submenu.dropdown-direction-up {
-  @apply top-auto bottom-full flex-col-reverse;
 }
 </style>
